fix: take ng-app and ng-controller apart from guest layout
the guest layout was designed at first to only be used at login route. However, it was later decided to reuse the layout at sign-up route as well. Thus, it was necessary to remove all logic related to login client-side logic from the layout. Both ng-app and ng-controller has been removed and put on the login view.