feat: add store subscription logic into store method in Subscription Controller
this commit introduces the algorithm behind the storage of new subscriptions.
The new method at first checks whether the request provides an opportunity
id or not. Afterwards, the method stores a new row in the subscriptions table
by calling the create static method of Subscription model. The next step
is to hit the '/subscriptions' endpoint on client side to give the user
the ability to apply to available opportunities.